requirements.txt
flask==3.0.2
gunicorn==21.2.0
streamlit==1.32.0
boto3==1.34.69
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
python-dotenv==1.0.1
